fix shareableListItemLoader unit test and add integration test
