+++ src/dataLoaders/shareableListItemLoader.spec.ts
@@ -1,91 +1,77 @@
 import { expect } from 'chai';
-import { PrismaClient } from '@prisma/client';
-import { client } from '../database/client';
-import { getShareableListItemsByItemIds } from './shareableListItemLoader';
-import {
-  clearDb,
-  createShareableListHelper,
-  createShareableListItemHelper,
-  createPilotUserHelper,
-} from '../test/helpers';
+import { faker } from '@faker-js/faker';
+import { ShareableListItem } from '../database/types';
+import { sortShareableListItemsByGivenItemIds } from './shareableListItemLoader';
 
-let db: PrismaClient;
-let list;
-let listItem;
-let itemIds = [];
-
-// this user will be put into the pilot
-const pilotUserHeaders = {
-  userId: '8009882300',
+const itemIds = [];
+const shareableListItemMaker = (itemId: string): ShareableListItem => {
+  return {
+    externalId: faker.datatype.uuid(),
+    itemId: parseInt(itemId) as unknown as bigint,
+    url: `${faker.internet.url()}/${faker.lorem.slug(5)}`,
+    title: faker.random.words(5),
+    excerpt: faker.lorem.sentences(2),
+    note: faker.lorem.sentences(2),
+    imageUrl: faker.image.cats(),
+    publisher: faker.company.name(),
+    authors: faker.name.fullName(),
+    sortOrder: 0,
+    createdAt: new Date(),
+    updatedAt: new Date(),
+  };
 };
 
 describe('shareableListItemLoader', () => {
-  beforeAll(async () => {
-    db = client();
-  });
-
-  afterAll(async () => {
-    await db.$disconnect();
-  });
-
-  beforeEach(async () => {
-    await clearDb(db);
-
-    // create pilot users
-    await createPilotUserHelper(db, {
-      userId: parseInt(pilotUserHeaders.userId),
-    });
-
-    // Create a List
-    list = await createShareableListHelper(db, {
-      userId: parseInt(pilotUserHeaders.userId),
-      title: 'Simon Le Bon List',
-    });
-    // Create some list items for that list
-    // then create some list items
-    const makeItems = Math.floor(Math.random() * 4) + 1;
-    itemIds = [];
-    for (let i = 0; i < makeItems; i++) {
-      listItem = await createShareableListItemHelper(db, {
-        list: list,
-      });
-      itemIds.push(parseInt(listItem.itemId as unknown as string));
-    }
-  });
-  describe('getShareableListItemsByItemIds', () => {
-    it('should return array of shareable list items by itemIds', async () => {
-      const listItems = await getShareableListItemsByItemIds(db, itemIds);
-      for (let i = 0; i < itemIds.length; i++) {
-        expect(parseInt(listItems[i].itemId as unknown as string)).to.equal(
-          itemIds[i]
-        );
-      }
-    });
-
-    it('should return empty array if itemIds do not exist', async () => {
-      const itemIds = [123, 543, 567];
-      const listItems = await getShareableListItemsByItemIds(
-        db,
-        itemIds as unknown as bigint[]
+  describe('sortShareableListItemsByGivenItemIds', () => {
+    it('should sort the shareable list items in the order of the given itemIds', () => {
+      // itemIds
+      const itemIds = [34, 12, 56, 78, 910];
+      // make some shareable list items matching the itemIds, making sure they are in a
+      // different order than the itemId array above
+      const shareableListItems: ShareableListItem[] = [
+        shareableListItemMaker('910n'),
+        shareableListItemMaker('78n'),
+        shareableListItemMaker('56n'),
+        shareableListItemMaker('12n'),
+        shareableListItemMaker('34n'),
+      ];
+      // they should now be sorted by the itemIds array
+      const sorted = sortShareableListItemsByGivenItemIds(
+        itemIds as unknown as bigint[],
+        shareableListItems
       );
-      expect(listItems.length).to.equal(0);
+
+      for (let i = 0; i < itemIds.length; i++) {
+        expect(sorted[i].itemId).to.equal(itemIds[i]);
+      }
     });
 
-    it('should return only list items that are found', async () => {
-      let listItems = await getShareableListItemsByItemIds(db, itemIds);
+    it('should return undefined in the place of itemId that was not found', () => {
+      // make some shareable list items matching some of the itemIdsn(one missing), making sure they are in a
+      // different order than the itemId array above
+      const shareableListItems: ShareableListItem[] = [
+        shareableListItemMaker('910n'),
+        shareableListItemMaker('78n'),
+        shareableListItemMaker('56n'),
+        shareableListItemMaker('34n'),
+      ];
+
+      // they should now be sorted by the itemIds array
+      const sorted = sortShareableListItemsByGivenItemIds(
+        itemIds as unknown as bigint[],
+        shareableListItems
+      );
+
+      // even though one itemId wasn't found the arrays should be of equal length
+      expect(sorted.length).to.equal(itemIds.length);
+
       for (let i = 0; i < itemIds.length; i++) {
-        expect(parseInt(listItems[i].itemId as unknown as string)).to.equal(
-          itemIds[i]
-        );
+        if (itemIds[i] !== '12n') {
+          expect(sorted[i].itemId).to.equal(itemIds[i]);
+        } else {
+          expect(sorted[i]).not.to.exist;
+        }
       }
-      // push another fake itemId
-      itemIds.push(5210346);
-      // request the array of list items again
-      listItems = await getShareableListItemsByItemIds(db, itemIds);
-      // lenghts should not match
-      expect(listItems.length).not.to.equal(itemIds.length);
-      // listItems length should be one less than itemIds length
-      expect(listItems.length).to.equal(itemIds.length - 1);
     });
   });
 });
